Refactor: Move Status as OutdatednessStatus to entities
