+++ lib/nanoc/base/entities.rb
@@ -20,4 +20,5 @@ require_relative 'entities/rule_memory'
 require_relative 'entities/site'
 require_relative 'entities/snapshot_def'
 
+require_relative 'entities/outdatedness_status'
 require_relative 'entities/dependency'
